+++ example/lib/main.dart
@@ -34,7 +34,7 @@ class _MyAppState extends State<MyApp> {
             ),
       },
       style: Style(
-        accentColor: Colors.green,
+        // accentColor: Colors.green,
       ),
     );
   }