+++ lib/fluent_ui.dart
@@ -17,6 +17,7 @@ export 'src/controls/inputs/checkbox.dart';
 export 'src/controls/inputs/drop_down_button.dart';
 export 'src/controls/inputs/icon_button.dart';
 export 'src/controls/inputs/radio_button.dart';
+export 'src/controls/inputs/slider.dart';
 export 'src/controls/inputs/split_button.dart';
 export 'src/controls/inputs/toggle_button.dart';
 export 'src/controls/inputs/toggle_switch.dart';